requirements.txt
# Deep-Live-Cam 依赖包清单

# 核心推理引擎
onnxruntime-gpu>=1.15.0  # GPU加速推理 (如果没有GPU可改为onnxruntime)
# onnxruntime>=1.15.0    # CPU版本 (备选)

# 计算机视觉
opencv-python>=4.8.0
# opencv-python-headless>=4.8.0  # 无GUI版本 (服务器环境)

# 人脸处理
insightface>=0.7.3
onnxruntime>=1.15.0  # InsightFace依赖

# 数值计算
numpy>=1.21.0

# 视频处理
ffmpeg-python>=0.2.0

# 进度显示和UI
tqdm>=4.64.0
rich>=13.0.0
PyQt5>=5.15.0  # 现代化GUI框架
Pillow>=10.0.0  # 图像处理支持

# 配置文件
pyyaml>=6.0

# 网络请求
requests>=2.28.0

# 图像处理增强 (可选)
# gfpgan>=1.3.8
# basicsr>=1.4.2
# facexlib>=0.3.0

# 开发和测试 (可选)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0

# 性能分析 (可选)
# psutil>=5.9.0
# py3nvml>=0.2.7

# Web界面 (可选)
# fastapi>=0.100.0
# uvicorn>=0.23.0
# streamlit>=1.25.0
